--- normalize_test.go
@@ -1,13 +1,0 @@
-package goalfred
-
-import "testing"
-
-func TestNormalize(t *testing.T) {
-	input := "äöü"
-	expected := "äöü"
-
-	normalized, err := Normalize(input)
-	if normalized != expected || err != nil {
-		t.Error("Expected", expected, "got", normalized)
-	}
-}
